Update 01_01_download_limits.R
I added the check and creation of folders to download the input data to make sure the script works, and have updated the link to the version 1.2 of the zip file. The one before was broken and this was pointed by the reviewer.
We could possibly keep the version 1.1 as before, but in this case we need to point to the correct file.